Support observation_delay of zero. Sanity check observation_delay. (#29)
In the differentiable version of control_arm_events, we need do the jax
analog of participants.shift(observation_delay). In the implementation,
doing this involved computing participants[:-observation_delay],
trimming the last few days off of the participants array. However, this
expression does the wrong thing with observation_delay is 0, returning
an empty array! This change fixes this bug.

Sanity checks on observation_delay have been added to the
non-differentiable version of control_arm_events.